Clean up for StaxTest:
* Added README-TEST.txt that indicates what StaxTest bundle contains,
  license (ASL 2.0 like ref. impl. itself)
* Bumped StaxTest version to 1.0rc1, so that although it'll be
  evolving over time, we can declare version 1.0 soon, if it seems
  useful and stable enough
* Added more author entries, needed for licensing (not all files have
  it, will update when there are actual changes to tests)

The main purpose is to clarify licensing, so that implementors feel
confident that they can use the test (esp. since it need not be included
with impl. distributions, just used as part of test or build process).

Thanks to Glenn Marcy for good suggestion on how to proceed with these
changes.